+++ bin/qds.py
@@ -583,10 +583,12 @@ def main():
                          help="cloud", choices=["AWS", "AZURE", "ORACLE_BMC", "ORACLE_OPC", "GCP"])
 
     optparser.add_option("--retry_delay", dest="retry_delay",
+                         type=int,
                          default=os.getenv('QDS_RETRY_DELAY'),
                          help="sleep interval between successive retries in case of retryable exceptions. defaults to 30s."
                         )
     optparser.add_option("--max_retries", dest="max_retries",
+                         type=int,
                          default=os.getenv('QDS_MAX_RETRIES'),
                          help="Number of re-attempts for an api-call in case of retryable exceptions. defaults to 6."
                         )
