+++ qds_sdk/connection.py
@@ -3,6 +3,7 @@ import requests
 import logging
 import ssl
 import json
+import time
 import pkg_resources
 from requests.adapters import HTTPAdapter
 from datetime import datetime
@@ -10,9 +11,8 @@ try:
     from requests.packages.urllib3.poolmanager import PoolManager
 except ImportError:
     from urllib3.poolmanager import PoolManager
-from qds_sdk.retry import retry
 from qds_sdk.exception import *
-from qds_sdk.qubole import Qubole
+from functools import wraps
 
 
 log = logging.getLogger("qds_connection")
@@ -35,7 +35,7 @@ class MyAdapter(HTTPAdapter):
 
 class Connection:
 
-    def __init__(self, auth, rest_url, skip_ssl_cert_check, reuse=True):
+    def __init__(self, auth, rest_url, skip_ssl_cert_check, reuse=True, max_retries=5, retry_delay=30):
         self.auth = auth
         self.rest_url = rest_url
         self.skip_ssl_cert_check = skip_ssl_cert_check
@@ -43,6 +43,8 @@ class Connection:
                          'Content-Type': 'application/json'}
 
         self.reuse = reuse
+        self.max_retries = max_retries
+        self.retry_delay = retry_delay
         if reuse:
             self.session = requests.Session()
             self.session.mount('https://', MyAdapter())
@@ -51,23 +53,42 @@ class Connection:
             self.session_with_retries = requests.Session()
             self.session_with_retries.mount('https://', MyAdapter(max_retries=3))
 
-    @retry((RetryWithDelay, requests.Timeout), tries=Qubole.max_retries, delay=Qubole.retry_delay, backoff=2)
+    def retry(ExceptionToCheck, tries=4, delay=3, backoff=2):
+        def deco_retry(f):
+            @wraps(f)
+            def f_retry(self,*args, **kwargs):
+                mtries, mdelay = self.max_retries, self.retry_delay
+                while mtries > 1:
+                    try:
+                        return f(self,*args, **kwargs)
+                    except ExceptionToCheck as e:
+                        logger = logging.getLogger("retry")
+                        msg = "%s, Retrying in %d seconds..." % (e.__class__.__name__, mdelay)
+                        logger.info(msg)
+                        time.sleep(mdelay)
+                        mtries -= 1
+                        mdelay *= backoff
+                return f(self,*args, **kwargs)
+            return f_retry  # true decorator
+        return deco_retry
+
+    @retry((RetryWithDelay, requests.Timeout))
     def get_raw(self, path, params=None):
         return self._api_call_raw("GET", path, params=params)
 
-    @retry((RetryWithDelay, requests.Timeout), tries=Qubole.max_retries, delay=Qubole.retry_delay, backoff=2)
+    @retry((RetryWithDelay, requests.Timeout))
     def get(self, path, params=None):
         return self._api_call("GET", path, params=params)
 
-    @retry((RetryWithDelay, requests.Timeout), tries=Qubole.max_retries, delay=Qubole.retry_delay, backoff=2)
+    @retry((RetryWithDelay, requests.Timeout))
     def put(self, path, data=None):
         return self._api_call("PUT", path, data)
 
-    @retry((RetryWithDelay, requests.Timeout), tries=Qubole.max_retries, delay=Qubole.retry_delay, backoff=2)
+    @retry((RetryWithDelay, requests.Timeout))
     def post(self, path, data=None):
         return self._api_call("POST", path, data)
 
-    @retry((RetryWithDelay, requests.Timeout), tries=Qubole.max_retries, delay=Qubole.retry_delay, backoff=2)
+    @retry((RetryWithDelay, requests.Timeout))
     def delete(self, path, data=None):
         return self._api_call("DELETE", path, data)
